stock/requirements.txt
ipywidgets
ydb
ydb[yc]
redis
python-binance
flask
#Flask-RESTX
#flask-swagger
flask-swagger-ui
python-dotenv
gate-api
loguru
pybit